requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
PyMuPDF==1.23.8
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
gunicorn==21.2.0